requirements.txt
# Core LangChain packages
langchain==0.2.16
langchain-openai==0.3.28
langchain-community==0.2.16
langchain-core>=0.2.40

# LangGraph for orchestration
langgraph==0.2.28
langsmith==0.1.117

# Vector database
qdrant-client==1.11.1
langchain-qdrant==0.2.0

# Web framework
streamlit==1.38.0

# PDF processing
PyPDF2==3.0.1
unstructured[pdf]==0.15.13
unstructured-client==0.26.0
pillow==10.4.0

# AI models and APIs
openai==1.51.0

# Evaluation (optional)
ragas==0.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2

# PubMed API integration
xmltodict==0.14.2

# Standard library backports for Python 3.12 compatibility
# (pathlib is built-in since Python 3.4)